MI5-58 Svaki @Composable sada ima svoju vlastitu .kt datoteku
Po zahtjevima iz #19
