+++ app/src/main/java/foi/air/szokpt/models/TransactionFilter.kt
@@ -1,0 +1,10 @@
+package foi.air.szokpt.models
+
+data class TransactionFilter(
+    val cardBrands: List<String>,
+    val trxTypes: List<String>,
+    val minAmount: String?,
+    val maxAmount: String?,
+    val afterDate: String?,
+    val beforeDate: String?
+)
