+++ app/src/main/java/foi/air/szokpt/views/app/TransactionsView.kt
@@ -1,7 +1,5 @@
 package foi.air.szokpt.views.app
 
-import android.widget.Toast
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
@@ -11,24 +9,14 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.layout.wrapContentHeight
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.rememberLazyListState
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.rounded.Add
-import androidx.compose.material3.Checkbox
-import androidx.compose.material3.CheckboxDefaults
 import androidx.compose.material3.CircularProgressIndicator
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.Icon
-import androidx.compose.material3.ModalBottomSheet
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.livedata.observeAsState
-import androidx.compose.runtime.mutableStateListOf
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
@@ -37,8 +25,6 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.vector.ImageVector
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.vectorResource
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
@@ -46,31 +32,15 @@ import androidx.compose.ui.unit.sp
 import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.navigation.NavController
 import foi.air.szokpt.R
-import foi.air.szokpt.ui.components.DatePickerField
-import foi.air.szokpt.ui.components.InputNumberField
+import foi.air.szokpt.models.TransactionFilter
+import foi.air.szokpt.ui.components.filter_components.ModalBottomSheetFilter
 import foi.air.szokpt.ui.components.interactible_components.BouncingFABDialogButton
-import foi.air.szokpt.ui.components.interactible_components.OutlineBouncingButton
 import foi.air.szokpt.ui.components.pagination_components.Pagination
+import foi.air.szokpt.ui.components.transaction_components.TransactionFilterView
 import foi.air.szokpt.ui.components.transaction_components.TransactionItem
-import foi.air.szokpt.ui.theme.BGLevelOne
-import foi.air.szokpt.ui.theme.Primary
 import foi.air.szokpt.ui.theme.TextWhite
-import foi.air.szokpt.ui.theme.danger
-import foi.air.szokpt.ui.theme.success
 import foi.air.szokpt.viewmodels.TransactionsViewModel
 import kotlinx.coroutines.launch
-import java.time.LocalDate
-import java.time.format.DateTimeFormatter
-
-
-data class FilterResults(
-    val cardBrands: List<String>,
-    val trxTypes: List<String>,
-    val minAmount: String?,
-    val maxAmount: String?,
-    val afterDate: String?,
-    val beforeDate: String?
-)
 
 @Composable
 fun TransactionsView(navController: NavController) {
@@ -88,7 +58,7 @@ fun TransactionsView(navController: NavController) {
     var isShowingFilters by remember { mutableStateOf(false) }
 
     // Result of applied filters. Contains a list of filters that are applied
-    var filterResults by remember { mutableStateOf<FilterResults?>(null) }
+    var filterResults by remember { mutableStateOf<TransactionFilter?>(null) }
 
     LaunchedEffect(currentPage) {
         if (transactionPage == null) {
@@ -181,7 +151,7 @@ fun TransactionsView(navController: NavController) {
             isExpanded = false
         },
         filterOptionsContent = {
-            TransactionFilter(
+            TransactionFilterView(
                 initialFilter = filterResults,
                 onApplyFilter = { results ->
                     filterResults = results
@@ -192,320 +162,4 @@ fun TransactionsView(navController: NavController) {
             )
         }
     )
-}
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun ModalBottomSheetFilter(
-    isVisible: Boolean,
-    onDismiss: () -> Unit,
-    hasFilters: Boolean,
-    isShowingFilters: Boolean,
-    onShowFilterOptions: () -> Unit,
-    onRemoveFilters: () -> Unit,
-    filterOptionsContent: @Composable () -> Unit
-) {
-    if (isVisible) {
-        ModalBottomSheet(
-            onDismissRequest = onDismiss,
-            containerColor = BGLevelOne,
-        ) {
-            LazyColumn(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .wrapContentHeight()
-                    .padding(16.dp),
-                verticalArrangement = Arrangement.spacedBy(5.dp),
-                horizontalAlignment = Alignment.CenterHorizontally
-            ) {
-                if (!isShowingFilters) {
-                    item {
-                        Row(
-                            modifier = Modifier
-                                .fillMaxWidth()
-                                .padding(9.dp)
-                                .clickable {
-                                    onRemoveFilters()
-                                },
-                            verticalAlignment = Alignment.CenterVertically
-                        ) {
-                            Icon(
-                                painter = painterResource(id = R.drawable.rounded_filter_alt_24),
-                                contentDescription = "Add Filters Icon",
-                                tint = success,
-                                modifier = Modifier.padding(end = 8.dp)
-                            )
-                            Text(
-                                text = if (hasFilters) "Change Current Filter" else "Apply New Filter",
-                                fontSize = 18.sp,
-                                color = success,
-                                modifier = Modifier
-                                    .clickable {
-                                        onShowFilterOptions()
-                                    }
-                            )
-                        }
-                    }
-
-                    if (hasFilters) {
-                        item {
-                            Row(
-                                modifier = Modifier
-                                    .fillMaxWidth()
-                                    .padding(9.dp)
-                                    .clickable {
-                                        onRemoveFilters()
-                                    },
-                                verticalAlignment = Alignment.CenterVertically
-                            ) {
-                                Icon(
-                                    painter = painterResource(id = R.drawable.rounded_filter_alt_off_24),
-                                    contentDescription = "Remove Filters Icon",
-                                    tint = danger,
-                                    modifier = Modifier.padding(end = 8.dp)
-                                )
-                                Text(
-                                    text = "Remove Filters",
-                                    fontSize = 18.sp,
-                                    color = danger
-                                )
-                            }
-                        }
-                    }
-                } else {
-                    item {
-                        filterOptionsContent()
-                    }
-                }
-            }
-        }
-    }
-}
-
-@Composable
-fun TransactionFilter(
-    initialFilter: FilterResults?,
-    onApplyFilter: (FilterResults) -> Unit
-) {
-    val cardBrands = listOf("Maestro", "Visa", "MasterCard", "Diners", "Discover")
-    val trxTypeMap = mapOf(
-        "sale" to "Sale",
-        "refund" to "Refund",
-        "void_sale" to "Void sale",
-        "void_refund" to "Void refund",
-        "reversal_sale" to "Reversal sale",
-        "reversal_refund" to "Reversal refund"
-    )
-
-    val selectedCardBrands = remember { mutableStateListOf<String>().apply { initialFilter?.cardBrands?.let { addAll(it) } } }
-    val selectedTrxTypes = remember { mutableStateListOf<String>().apply { initialFilter?.trxTypes?.let { addAll(it) } } }
-
-    var minAmount by remember { mutableStateOf(initialFilter?.minAmount) }
-    var maxAmount by remember { mutableStateOf(initialFilter?.maxAmount) }
-    fun isValidAmountRange(min: String?, max: String?): Boolean {
-        val minValue = min?.toDoubleOrNull()
-        val maxValue = max?.toDoubleOrNull()
-        return minValue == null || maxValue == null || minValue <= maxValue
-    }
-
-    val dateFormatter = DateTimeFormatter.ofPattern("dd/MM/yyyy")
-    var selectedAfterDate by remember {
-        mutableStateOf(
-            initialFilter?.afterDate?.let { LocalDate.parse(it, dateFormatter) }
-        )
-    }
-    var selectedBeforeDate by remember {
-        mutableStateOf(
-            initialFilter?.beforeDate?.let { LocalDate.parse(it, dateFormatter) }
-        )
-    }
-    fun validateAndSetAfterDate(date: LocalDate) {
-        if (selectedBeforeDate == null || date.isBefore(selectedBeforeDate) || date.isEqual(selectedBeforeDate)) {
-            selectedAfterDate = date
-        }
-    }
-    fun validateAndSetBeforeDate(date: LocalDate) {
-        if (selectedAfterDate == null || date.isAfter(selectedAfterDate) || date.isEqual(selectedAfterDate)) {
-            selectedBeforeDate = date
-        }
-    }
-
-    Column(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(8.dp)
-    ) {
-        Column(
-            verticalArrangement = Arrangement.spacedBy(16.dp)
-        ) {
-            Row(
-                modifier = Modifier.fillMaxWidth(),
-                horizontalArrangement = Arrangement.SpaceEvenly
-            ) {
-                Column(
-                    modifier = Modifier.weight(1f),
-                    verticalArrangement = Arrangement.spacedBy(8.dp)
-                ) {
-                    Text(
-                        text = "Transaction types",
-                        fontSize = 18.sp,
-                        fontWeight = FontWeight.Medium
-                    )
-
-                    trxTypeMap.forEach { (key, displayValue) ->
-                        Row(
-                            verticalAlignment = Alignment.CenterVertically,
-                            modifier = Modifier
-                                .fillMaxWidth()
-                                .clickable {
-                                    if (selectedTrxTypes.contains(key)) {
-                                        selectedTrxTypes.remove(key)
-                                    } else {
-                                        selectedTrxTypes.add(key)
-                                    }
-                                }
-                                .padding(vertical = 4.dp)
-                        ) {
-                            Checkbox(
-                                checked = selectedTrxTypes.contains(key),
-                                onCheckedChange = { isChecked ->
-                                    if (isChecked) selectedTrxTypes.add(key) else selectedTrxTypes.remove(key)
-                                },
-                                colors = CheckboxDefaults.colors(
-                                    checkedColor = Primary
-                                )
-                            )
-                            Text(
-                                text = displayValue,
-                                fontSize = 14.sp,
-                                modifier = Modifier.padding(start = 8.dp)
-                            )
-                        }
-                    }
-                }
-
-                Spacer(modifier = Modifier.width(16.dp))
-
-                Column(
-                    modifier = Modifier.weight(1f),
-                    verticalArrangement = Arrangement.spacedBy(8.dp)
-                ) {
-                    Text(
-                        text = "Card brands",
-                        fontSize = 18.sp,
-                        fontWeight = FontWeight.Medium
-                    )
-
-                    cardBrands.forEach { brand ->
-                        Row(
-                            verticalAlignment = Alignment.CenterVertically,
-                            modifier = Modifier
-                                .fillMaxWidth()
-                                .clickable {
-                                    if (selectedCardBrands.contains(brand)) {
-                                        selectedCardBrands.remove(brand)
-                                    } else {
-                                        selectedCardBrands.add(brand)
-                                    }
-                                }
-                                .padding(vertical = 4.dp)
-                        ) {
-                            Checkbox(
-                                checked = selectedCardBrands.contains(brand),
-                                onCheckedChange = { isChecked ->
-                                    if (isChecked) selectedCardBrands.add(brand) else selectedCardBrands.remove(brand)
-                                },
-                                colors = CheckboxDefaults.colors(
-                                    checkedColor = Primary
-                                )
-                            )
-                            Text(
-                                text = brand,
-                                fontSize = 14.sp,
-                                modifier = Modifier.padding(start = 8.dp)
-                            )
-                        }
-                    }
-                }
-            }
-        }
-
-        Text(
-            text = "Date range",
-            fontSize = 18.sp,
-            fontWeight = FontWeight.Medium,
-        )
-        Row(
-            modifier = Modifier.fillMaxWidth(),
-            horizontalArrangement = Arrangement.SpaceEvenly,
-        ) {
-
-            DatePickerField(
-                onDateSelected = { date -> validateAndSetAfterDate(date) },
-                label = "After date",
-                initialDate = selectedAfterDate,
-                maxWidth = 172.dp
-            )
-            DatePickerField(
-                onDateSelected = { date -> validateAndSetBeforeDate(date) },
-                label = "Before date",
-                initialDate = selectedBeforeDate,
-                maxWidth = 172.dp
-            )
-        }
-
-        Text(
-            text = "Amount value range",
-            fontSize = 18.sp,
-            fontWeight = FontWeight.Medium,
-        )
-        Row(
-            modifier = Modifier.fillMaxWidth(),
-            horizontalArrangement = Arrangement.SpaceEvenly,
-        ) {
-            InputNumberField(
-                label = "Min",
-                value = minAmount ?: "", // Provide an empty string if null
-                onValueChange = { minAmount = it.takeIf { it.isNotEmpty() } },
-                width = 160.dp,
-            )
-            InputNumberField(
-                label = "Max",
-                value = maxAmount ?: "", // Provide an empty string if null
-                onValueChange = { maxAmount = it.takeIf { it.isNotEmpty() } },
-                width = 160.dp,
-            )
-        }
-
-        Spacer(modifier = Modifier.height(8.dp))
-        val context = LocalContext.current
-        Row(
-            modifier = Modifier.fillMaxWidth().padding(bottom = 16.dp),
-            horizontalArrangement = Arrangement.Center,
-        ) {
-            OutlineBouncingButton(
-                inputText = "Apply filter",
-                inputIcon = Icons.Rounded.Add,
-                contentColor = success,
-                borderColor = success,
-                onClick = {
-
-                    if (isValidAmountRange(minAmount, maxAmount)) {
-                        val results = FilterResults(
-                            cardBrands = selectedCardBrands.toList(),
-                            trxTypes = selectedTrxTypes.toList(),
-                            minAmount = minAmount?.takeIf { it.isNotEmpty() },
-                            maxAmount = maxAmount?.takeIf { it.isNotEmpty() },
-                            afterDate = selectedAfterDate?.format(dateFormatter),
-                            beforeDate = selectedBeforeDate?.format(dateFormatter)
-                        )
-                        onApplyFilter(results)
-                    } else {
-                        println("Invalid amount range: min=$minAmount, max=$maxAmount")
-                        Toast.makeText(context, "Invalid amount range: Min must be less than or equal to Max", Toast.LENGTH_SHORT).show()
-                    }
-                },
-            )
-        }
-    }
 }
